+++ example/realworld.10000.js
@@ -1,7 +1,7 @@
 //An extract of address points from the LINZ bulk extract: http://www.linz.govt.nz/survey-titles/landonline-data/landonline-bde
 //Should be this data set: http://data.linz.govt.nz/#/layer/779-nz-street-address-electoral/
 export const addressPoints = [
-[-37.8839, 175.3745188667, "571"],
+[-37.8839, null, "571"],
 [-37.8869090667, 175.3657417333, "486"],
 [-37.8894207167, 175.4015351167, "807"],
 [-37.8927369333, 175.4087452333, "899"],