+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.1.0 Release
+
+- Handle invalid points gracefully
+- Add `fitBoundsOnUpdate` prop to indicate that the map should fit the data in the map on update
+
 # 0.0.2 Release
 
 - Fix an issue with the gradient property being applied